refactor: reduce LOC in large files by extracting components
Extract focused classes from large files to improve maintainability:

- storage_factory.py (617→155 LOC): Extract StorageAdapter to storage_adapter.py
- chembl/client.py (551→505 LOC): Extract ChemblEntityMapper to entity_mapper.py
- generic_factory.py (413→360 LOC): Extract ServicesBuilder to services_builder.py
- pubmed_client.py (353→336 LOC): Extract PubMedXmlProcessor to xml_processor.py
- uniprot/client.py (319→296 LOC): Extract FastaParser to fasta_parser.py

Total net reduction: ~300 LOC from target files
New extracted modules: 5 files (~500 LOC focused components)
